# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.InternalErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Internal server error
        type: string
    type: object
...
  handlers.InvalidDataErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid data
        type: string
    type: object
  handlers.InvalidDataOrTokenErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid token or data
        type: string
    type: object
...
  handlers.InvalidParametersErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
  handlers.InvalidTokenErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid token
        type: string
    type: object
...
  handlers.MaterialsResponse:
    description: Ответ с результатами поиска материалов
    properties:
      hasMore:
        example: true
        type: boolean
      limit:
        example: 20
        type: integer
      materials:
        items:
          $ref: '#/definitions/models.CatalogMaterial'
        type: array
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
    type: object
...
  handlers.RefreshTokenRequest:
    description: Запрос на обновление токенов
    properties:
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refreshToken
    type: object
...
  handlers.SubjectsResponse:
    description: Ответ со списком предметов
    properties:
      subjects:
        items:
          $ref: '#/definitions/models.Subject'
        type: array
    type: object
  handlers.SuccessResponse:
    description: Стандартный успешный ответ
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
  handlers.TeachersResponse:
    description: Ответ с результатами поиска преподавателей
    properties:
      teachers:
        items:
          $ref: '#/definitions/models.Teacher'
        type: array
    type: object
...
  models.AuthResponse:
    description: Ответ с токенами и данными пользователя
    properties:
      accessToken:
        type: string
      message:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Author:
    description: Автор материала
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Иван Иванов
        type: string
    type: object
...
  models.CatalogMaterial:
    description: Материал в каталоге
    properties:
      author:
        $ref: '#/definitions/models.Author'
      duration:
        example: 120
        type: integer
      id:
        example: 1
        type: integer
      level:
        example: beginner
        type: string
      rating:
        example: 4.8
        type: number
      studentsCount:
        example: 150
        type: integer
      subject:
        example: math
        type: string
      thumbnailUrl:
        example: https://example.com/thumbnail.jpg
        type: string
      title:
        example: Основы алгебры
        type: string
    type: object
...
  models.ForgotPasswordRequest:
    description: Запрос на сброс пароля
    properties:
      email:
        type: string
    required:
    - email
    type: object
  models.LoginRequest:
    description: Запрос на вход
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.RegisterRequest:
    description: Запрос на регистрацию
    properties:
      email:
        type: string
      fullName:
        type: string
      password:
        minLength: 6
        type: string
      role:
        enum:
        - student
        - teacher
        - admin
        type: string
    required:
    - email
    - fullName
    - password
    - role
    type: object
  models.ResetPasswordRequest:
    description: Запрос на установку нового пароля
    properties:
      newPassword:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - newPassword
    - token
    type: object
...
  models.Subject:
    description: Учебный предмет
    properties:
      icon:
        example: "\U0001F4D0"
        type: string
      id:
        example: math
        type: string
      name:
        example: Математика
        type: string
    type: object
  models.Teacher:
    description: Преподаватель в каталоге
    properties:
      avatarUrl:
        example: https://example.com/avatar.jpg
        type: string
      id:
        example: 1
        type: integer
      materialsCount:
        example: 25
        type: integer
      name:
        example: Мария Петрова
        type: string
      rating:
        example: 4.9
        type: number
      specializations:
        example:
        - math
        - physics
        items:
          type: string
        type: array
    type: object
...
  models.User:
    description: Модель пользователя
    properties:
      avatarUrl:
        type: string
      blockReason:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: integer
      isBlocked:
        type: boolean
      isVerified:
        type: boolean
      role:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для образовательной платформы Пайдея
  title: Paydeya Education Platform API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Отправляет инструкции по сбросу пароля на email
      parameters:
      - description: Email для сброса пароля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Инструкции отправлены
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      summary: Запрос сброса пароля
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и возвращает токены
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/handlers.InvalidDataErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Завершает сеанс пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет access и refresh токены
      parameters:
      - description: Refresh токен
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токены обновлены
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "401":
          description: Невалидный токен
          schema:
            $ref: '#/definitions/handlers.InvalidTokenErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      summary: Обновление токенов
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает токены
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Устанавливает новый пароль по токену сброса
      parameters:
      - description: Данные для сброса пароля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пароль изменен
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверные данные или токен
          schema:
            $ref: '#/definitions/handlers.InvalidDataOrTokenErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      summary: Сброс пароля
      tags:
      - auth
  /catalog/materials:
    get:
      consumes:
      - application/json
      description: Возвращает материалы с фильтрацией и пагинацией
      parameters:
      - description: Поисковый запрос
        in: query
        name: search
        type: string
      - description: Фильтр по предмету
        in: query
        name: subject
        type: string
      - description: Фильтр по уровню сложности
        in: query
        name: level
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Количество материалов на странице
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список материалов
          schema:
            $ref: '#/definitions/handlers.MaterialsResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      summary: Поиск материалов в каталоге
      tags:
      - catalog
  /catalog/subjects:
    get:
      consumes:
      - application/json
      description: Возвращает все доступные учебные предметы
      produces:
      - application/json
      responses:
        "200":
          description: Список предметов
          schema:
            $ref: '#/definitions/handlers.SubjectsResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      summary: Получить список предметов
      tags:
      - catalog
  /catalog/teachers:
    get:
      consumes:
      - application/json
      description: Возвращает преподавателей с фильтрацией
      parameters:
      - description: Поисковый запрос
        in: query
        name: search
        type: string
      - description: Фильтр по предмету
        in: query
        name: subject
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список преподавателей
          schema:
            $ref: '#/definitions/handlers.TeachersResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      summary: Поиск преподавателей
      tags:
      - catalog
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...